+++ taglib/trueaudio/trueaudioproperties.cpp
@@ -145,7 +145,7 @@ void TrueAudio::Properties::read(const ByteVector &data, offset_t streamLength)
     if(d->sampleFrames > 0 && d->sampleRate > 0) {
       const auto length = static_cast<double>(d->sampleFrames) * 1000.0 / d->sampleRate;
       d->length  = static_cast<int>(length + 0.5);
-      d->bitrate = static_cast<int>(streamLength * 8.0 / length + 0.5);
+      d->bitrate = static_cast<int>(static_cast<double>(streamLength) * 8.0 / length + 0.5);
     }
   }
 }
